requirements/local.txt
# Local (development) dependencies.
-r base.txt

coverage>=3.7.1
django-debug-toolbar>=1.2.1
